Added first working GUI
This is the first prototype GUI for use the three webtext script.
It is functional and has the option to send a copy text to the user
who sends the message.

Might need to make it like a desktop app or something like to make
it usable by people who don't code